# Pull threading.local out of the generated repr's globals.