--- a/src/attr/_compat.py
+++ b/src/attr/_compat.py
@@ -2,6 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 import platform
 import sys
+import threading
 import types
 import warnings
 
@@ -243,3 +244,17 @@ def make_set_closure_cell():
 
 
 set_closure_cell = make_set_closure_cell()
+
+# Thread-local global to track attrs instances which are already being repr'd.
+# This is needed because there is no other (thread-safe) way to pass info
+# about the instances that are already being repr'd through the call stack
+# in order to ensure we don't perform infinite recursion.
+#
+# For instance, if an instance contains a dict which contains that instance,
+# we need to know that we're already repr'ing the outside instance from within
+# the dict's repr() call.
+#
+# This lives here rather than in _make.py so that the functions in _make.py
+# don't have a direct reference to the thread-local in their globals dict.
+# If they have such a reference, it breaks cloudpickle.
+repr_context = threading.local()