@Tinche 2021-11-01
@thetorpedodog  Could you show us the performance impact of this change on a sample class? I recommend `pyperf`.