@thetorpedodog 2021-11-01
```
import textwrap

import attr
import pyperf

@attr.s
class Foo(object):
    bar = attr.ib()
    baz = attr.ib()
    boz = attr.ib()


runner = pyperf.Runner()
runner.timeit(
    name="basic repr",
    stmt="repr(x)",
    setup=textwrap.dedent("""
        import __main__
        x = __main__.Foo(1, "two", 3.0)
    """),
)

runner.timeit(
    name="recursion repr",
    stmt="repr(x)",
    setup=textwrap.dedent("""
        import __main__
        x = __main__.Foo(1, "two", [])
        x.boz.extend((x, x))
    """),
)
```

Benchmarks from Python 3.9.2 on Debian:

Before:

```
(venv) pfish@tiledebian:~/attrs$ python ./microbench.py 
.....................
basic repr: Mean +- std dev: 674 ns +- 10 ns
.....................
recursion repr: Mean +- std dev: 1.02 us +- 0.02 us
```

After:

```
(venv) pfish@tiledebian:~/attrs$ python ./microbench.py 
.....................
basic repr: Mean +- std dev: 673 ns +- 12 ns
.....................
recursion repr: Mean +- std dev: 1.02 us +- 0.02 us
```

Obviously this means it is 1 nsec (± 22 nsec) faster, a clear and meaningful performance improvement that is not at all probably just noise :)