@Tinche 2021-11-01
@thetorpedodog Interesting, because that's not what I'm seeing at all ;)

`main` branch, now:
```
> pyperf timeit -g -s "from attr import make_class, attrib; A = make_class('A', ['a', 'b', 'c', 'd', 'e']); a = A(1, 2, 3, 4, 5)" "repr(a)"
.....................
 916 ns: 1 #########
 923 ns: 1 #########
 930 ns: 2 ##################
 937 ns: 2 ##################
 944 ns: 6 #####################################################
 951 ns: 6 #####################################################
 958 ns: 7 #############################################################
 965 ns: 5 ############################################
 972 ns: 9 ###############################################################################
 979 ns: 6 #####################################################
 986 ns: 7 #############################################################
 993 ns: 3 ##########################
 999 ns: 3 ##########################
1.01 us: 0 |
1.01 us: 1 #########
1.02 us: 0 |
1.03 us: 0 |
1.03 us: 0 |
1.04 us: 0 |
1.05 us: 0 |
1.05 us: 1 #########

Mean +- std dev: 970 ns +- 23 ns
```
Your branch:
```
> pyperf timeit -g -s "from attr import make_class, attrib; A = make_class('A', ['a', 'b', 'c', 'd', 'e']); a = A(1, 2, 3, 4, 5)" "repr(a)"
.....................
2.27 us: 1 ##########
2.28 us: 1 ##########
2.29 us: 0 |
2.30 us: 1 ##########
2.31 us: 1 ##########
2.32 us: 1 ##########
2.33 us: 2 ####################
2.34 us: 3 ##############################
2.35 us: 2 ####################
2.36 us: 5 #################################################
2.37 us: 5 #################################################
2.38 us: 8 ###############################################################################
2.39 us: 5 #################################################
2.41 us: 7 #####################################################################
2.42 us: 4 ########################################
2.43 us: 3 ##############################
2.44 us: 3 ##############################
2.45 us: 4 ########################################
2.46 us: 1 ##########
2.47 us: 2 ####################
2.48 us: 1 ##########

Mean +- std dev: 2.39 us +- 0.05 us
```
(CPython 3.9.7 on Ubuntu.)

That's a very significant slowdown. My gut feeling was there was no chance the proposed code would be as fast as the code that's in there now since not only are functions calls kinda expensive in CPython, you're using a context manager too.