@Tinche 2021-11-01
Your script shows the same slowdown btw. I suspect you've forgotten to switch the branch between runs. Sorry :)