@hynek 2021-11-02
TBH I wouldn't consider performance of repr a top priority. Isn't it usually used only in interactive use?

My gut feeling is that cloudpickle interaction is a bigger problem because it's widely used in the data science sphere. 🤔