@Tinche 2021-11-02
I `repr` on every request, to log the request description. Why should I pay the price for a library I've never used? *shrug*