@thetorpedodog 2021-11-02
> I `repr` on every request, to log the request description. Why should I pay the price for a library I've never used? *shrug*

I get the concern here, but if there were code where repr was the hottest path and the thing most limiting performance, that would be a sign of rather deeper architectural problems. My feeling about having a special switch in the decorator for this is that it would be an excessively special case that would be useful only in a vanishingly small number of cases, and at that point the author would probably want to hand-write the Python themselves.

> That said, I don't understand the exact problem here. Cloudpickle doesn't like threadlocals in the globals of `__repr__`, but having `__repr__` call a function that does have a threadlocal in globals is fine for some reason?

It’s, yeah, a bit confusing. I think the deal is that cloudpickle will try to pickle everything in a class instance itself by value, but once it goes outside that class, it will pickle by reference. Or something like that.

The benefit of this is that we can limit the performance impact of this change by restructuring it: simply [pull the threadlocal into another module](https://github.com/thetorpedodog/attrs/tree/pull-it-into-a-module), then reference only that *module* in the `__repr__`’s globals. (I used the `_compat` module, because, uhhhhhh, it’s for compatibility with cloudpickle? Does that make sense? Any better ideas?) That change is sitting on another branch: https://github.com/thetorpedodog/attrs/commit/0d710b451f93b34481a62b50759576ef943e8e0c

For comparison (and this time I **didn’t** screw it up):

Old:

```plain
(a-completely-different-venv) pfish@tiledebian:~/a-completely-different-folder$ python ./microbench.py
.....................
basic repr: Mean +- std dev: 654 ns +- 16 ns
.....................
recursion repr: Mean +- std dev: 988 ns +- 23 ns
```

Thread-local in a different module (0d710b451f93b34481a62b50759576ef943e8e0c):

```plain
(venv) pfish@tiledebian:~/attrs$ python ./microbench.py
.....................
basic repr: Mean +- std dev: 682 ns +- 16 ns
.....................
recursion repr: Mean +- std dev: 1.05 us +- 0.02 us
```

Context-manager (2063781585ecec4f31c4efe6f4fc78097752bb65):

```plain
(venv) pfish@tiledebian:~/attrs$ python ./microbench.py
.....................
basic repr: Mean +- std dev: 1.67 us +- 0.05 us
.....................
recursion repr: Mean +- std dev: 3.77 us +- 0.07 us
```

I think that cost is low enough to merit not worrying about given that it fixes real user issues.