@Tinche 2021-11-02
That does look much better. Can we add a unit test for this as well (adding cloudpickle to the dev dependencies)?