@thetorpedodog 2021-11-02
Switched this branch over to point at the version that pulls the thread-local into `_compat`. The previous version (context manager) is still available on my [`manage-the-context` branch](https://github.com/thetorpedodog/attrs/tree/manage-the-context).